+++ game1/module/src/main/java/com/example/module/service/TypeService.java
@@ -1,7 +1,7 @@
 package com.example.module.service;
 
-import com.example.module.entity.Game;
 import com.example.module.entity.Type;
+import com.example.module.mapper.GameMapper;
 import com.example.module.mapper.TypeMapper;
 import jakarta.annotation.Resource;
 import org.springframework.stereotype.Service;
@@ -13,6 +13,9 @@ import java.util.List;
 public class TypeService {
     @Resource
     private TypeMapper mapper;
+    @Resource
+    private GameMapper gameMapper;
+
     public Type getById(BigInteger id) {
         return mapper.getById(id);
     }
@@ -33,6 +36,9 @@ public class TypeService {
     public int delete(BigInteger id) {
         if (id == null){
             throw new RuntimeException("id 不能为空");
+        }
+        if (gameMapper.isExistByTypeId(id)!=null){
+            throw new RuntimeException("该类型下存在游戏，不能删除");
         }
         int time = (int) (System.currentTimeMillis() / 1000);
         return mapper.delete(id, time);
